+++ src/client/containers/Admin-Event-details/Admin-Event-modal-view.js
@@ -2,31 +2,17 @@ import React from 'react';
 import { EventCardText } from '../../components/EventCardText/EventCardText';
 import Modal from '../../components/Modal/Modal';
 import { StudyGroupLogo } from '../../components/StudyGroupLogo/StudyGroupLogo';
-import PropTypes from 'prop-types';
+import PropTypes, { shape } from 'prop-types';
 import Elasticlogo from './Elastic.logo.svg';
 import './Admin-event-modal.style.css';
 
-export default function EventModalView(props) {
-  const students = [
-    {
-      studentName: 'DemoStudent1',
-      class: 'class 15',
-      topic: 'Topic: JavaScript week3 homework,classes,Function',
-      code: 'maherweb.dk',
-    },
-    {
-      studentName: 'DemoStudent2',
-      class: 'class 15',
-      topic: 'Topic: JavaScript week3 homework,classes,Function',
-      code: 'maherweb.dk',
-    },
-    {
-      studentName: 'DemoStudent3',
-      class: 'class 15',
-      topic: 'Topic: JavaScript week3 homework,classes,Function',
-      code: 'maherweb.dk',
-    },
-  ];
+export default function EventModalView({
+  eventTitle,
+  eventDate,
+  mentor,
+  eventLink,
+  items,
+}) {
   return (
     <div className="event-modal-view">
       <Modal>
@@ -36,46 +22,53 @@ export default function EventModalView(props) {
           </div>
           <div className="card-text">
             <EventCardText
-              title={props.eventTitle}
-              datetime={props.eventDate}
-              mentor={props.mentor}
-              link={props.eventLink}
+              title={eventTitle}
+              datetime={eventDate}
+              mentor={mentor}
+              link={eventLink}
             />
           </div>
         </div>
         <div className="student-details-block">
-          {students.map((elem) => {
-            return (
-              <div className="student-info">
-                <p>
-                  {elem.studentName} | {elem.class}
-                  <span className="close-btn">x</span>
-                </p>
-
-                <p>Topic: {elem.topic}</p>
-                <p>
-                  code :
-                  <a href={elem.code} className="student-code-link">
-                    {elem.code}
-                  </a>
-                </p>
-              </div>
-            );
-          })}
+          {items.map((student) => (
+            <div className="student-info">
+              {student.studentName} | {student.class}
+              <p>
+                <span className="close-btn">x</span>
+              </p>
+              <p>Topic: {student.topic}</p>
+              <p>
+                code :
+                <a href={student.code} className="student-code-link">
+                  {student.code}
+                </a>
+              </p>
+            </div>
+          ))}
         </div>
       </Modal>
     </div>
   );
 }
+
 EventModalView.propTypes = {
   eventTitle: PropTypes.string,
   eventDate: PropTypes.string,
   mentor: PropTypes.string,
   eventLink: PropTypes.string,
+  items: PropTypes.arrayOf(
+    shape({
+      studentName: PropTypes.string,
+      class: PropTypes.string,
+      topic: PropTypes.string,
+      code: PropTypes.string,
+    }),
+  ),
 };
 EventModalView.defaultProps = {
   eventTitle: '',
   eventDate: '',
   mentor: '',
   eventLink: '',
+  items: [],
 };
